Change to use Dashboard skeleton for placeholder
A loading skeleton is a simplified version of the UI.

Many websites use them as a placeholder (or fallback) to indicate to users that the content is loading. Any UI you embed into loading.tsx will be embedded as part of the static file, and sent first. Then, the rest of the dynamic content will be streamed from the server to the client.